Refactor creating expiry DateTime objects into a reusable function
The expiry DateTime objects ($expireDate and $oldDate) are both
initialised in exactly the same way. Given that, it makes sense to
refactor them into a reusable utility method.

Doing so makes the code more compact and easier to read. The commit also
removes the possibility of throwing a silent error, as previously the
DateTime modify method could be called, even if there was no expireDate
provided to the command.