Refactor magic number to be a class constant
Like the two other refactors that have done similar things, this commit
replaces the reference to a magic number with a meaningfully named
constant. Perhaps it's not necessary, but my thinking is: "Would a new
developer necessarily know that 0 means success?". So it seems better
not to have to ask the question in the first place.